+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Switched config to use Forge config system
+
 ## [1.3.1] - 2020-12-17
 ### Changed
 - Add `killed_by_player` condition to player entity group loot table